requirements.txt
pandas<2.0.0,>=1.4.4
requests>=2.3.0
requests-cache>=0.5.0
python-decouple
aiohttp